Extract reusable LLM labeler service
