+++ vaannotate/vaannotate_ai_backend/core/__init__.py
@@ -1,0 +1,20 @@
+"""Core primitives for the VAAnnotate AI backend."""
+
+from .data import DataRepository
+from .embeddings import (
+    EmbeddingStore,
+    IndexConfig,
+    Models,
+    build_models_from_env,
+)
+from .retrieval import RetrievalCoordinator, SemanticQuery
+
+__all__ = [
+    "DataRepository",
+    "EmbeddingStore",
+    "IndexConfig",
+    "Models",
+    "build_models_from_env",
+    "RetrievalCoordinator",
+    "SemanticQuery",
+]
