+++ vaannotate/vaannotate_ai_backend/core/retrieval.py
@@ -1,0 +1,5 @@
+"""Retrieval coordination primitives for combining semantic and keyword search."""
+
+from ..retrieval_coordinator import RetrievalCoordinator, SemanticQuery
+
+__all__ = ["RetrievalCoordinator", "SemanticQuery"]
