+++ vaannotate/vaannotate_ai_backend/services/__init__.py
@@ -1,0 +1,6 @@
+"""Service-layer utilities for active learning and inference pipelines."""
+
+from .context_builder import ContextBuilder
+from .llm_labeler import LLMLabeler, LLM_RECORDER
+
+__all__ = ["ContextBuilder", "LLMLabeler", "LLM_RECORDER"]
